Move FnRef details into a __private header
