+++ subspace/CMakeLists.txt
@@ -45,7 +45,9 @@ target_sources(subspace PUBLIC
     "containers/array.h"
     "containers/slice.h"
     "containers/vec.h"
+    "fn/__private/callable_types.h"
     "fn/__private/fn_box_storage.h"
+    "fn/__private/fn_ref_invoker.h"
     "fn/__private/signature.h"
     "fn/callable.h"
     "fn/fn.h"
